Extract NumberConsumer
