+++ lib/javascript/tokenizer/grammar.rb
@@ -1,3 +1,5 @@
+require "javascript/tokenizer/grammar/number_consumer"
+
 module Javascript
   class Grammar
     START_OF_IDENTIFIER = /\p{L}|_|\$/
@@ -95,67 +97,7 @@ module Javascript
       end
 
       def consume_number
-        case
-        when scanner.scan(/0x/i)     then consume_nondecimal_number(base: 16)
-        when scanner.scan(/0b/i)     then consume_nondecimal_number(base: 2)
-        when scanner.scan(/0o/i)     then consume_nondecimal_number(base: 8)
-        when scanner.scan(/0(?=\d)/) then consume_potentially_nondecimal_number(base: 8)
-        else                              consume_decimal_number
-        end
-      end
-
-      def consume_nondecimal_number(base:)
-        consume_integer(pattern: /#{Regexp.union((Array("0".."9") + Array("a".."z"))[0...base]).source}/i).to_i(base)
-      end
-
-      def consume_potentially_nondecimal_number(base:)
-        decimal = consume_decimal_number
-
-        if decimal.to_i == decimal
-          Integer(decimal.to_i.to_s, base, exception: false) || decimal
-        else
-          decimal
-        end
-      end
-
-      def consume_decimal_number
-        if scanner.scan(".")
-          whole      = 0
-          fractional = consume_integer
-        else
-          whole      = consume_integer
-          fractional = scanner.scan(".") && scanner.peek(1).match?(/\d/) ? consume_integer : 0
-        end
-
-        exponent = scanner.scan(/e/i) ? consume_signed_integer : 0
-
-        "#{whole}.#{fractional}e#{exponent}".to_f
-      end
-
-      def consume_integer(pattern: /\d/)
-        [].tap do |digits|
-          loop do
-            case
-            when scanner.scan(pattern)
-              digits << scanner.matched
-            when scanner.scan("_")
-              raise SyntaxError unless digits.last&.match?(pattern) && scanner.peek(1).match?(pattern)
-            else
-              if digits.none?
-                raise SyntaxError
-              else
-                break
-              end
-            end
-          end
-        end.join
-      end
-
-      def consume_signed_integer(pattern: /\d/)
-        sign    = scanner.scan(/\+|\-/).to_s
-        integer = consume_integer(pattern: pattern)
-
-        sign + integer
+        NumberConsumer.new(scanner).consume_number
       end
 
       def tokenize_string
